requirements.txt
backtrader
pandas
PyYAML
requests
httpx
scikit-learn
joblib
fastapi
uvicorn[standard]
